cookieのJWT関連の設定

+++ backend/api/accounts/views.py
@@ -1,10 +1,11 @@
 from django.contrib.auth import get_user_model
 from rest_framework.generics import ListAPIView
 from .serializers import CustomUserSerializer
+from .authentication import CookieJWTAuthentication
 from rest_framework_simplejwt import views
 from rest_framework_simplejwt import exceptions
 from rest_framework.response import Response
-from rest_framework import status
+from rest_framework import status, permissions
 
 User = get_user_model()
 
@@ -25,17 +26,12 @@ class JWTokenObtainView(views.TokenObtainPairView):
         
         response = Response(serializer.validated_data, status=status.HTTP_200_OK)
 
-        try:
-            response.delete_cookie("access_token")
-        except Exception as e:
-            print(e)
-        
-        # Cookieのheaderにトークンのセット
+        # Cookieにトークンをセット
         response.set_cookie(
             "access_token",
             serializer.validated_data["access"],
-            # 期限は1日
-            max_age=60 * 60 * 24,
+            # 期限は3時間
+            max_age=60 * 60 * 3,
             httponly=True,
         )
         response.set_cookie(
@@ -52,6 +48,14 @@ class JWTokenObtainView(views.TokenObtainPairView):
 # JWTのリフレッシュ
 class JWTokenRefreshView(views.TokenRefreshView):
     def post(self, request, *args, **kwargs):
+        # cookieからリフレッシュトークンを取得
+        refresh_token = request.COOKIES.get("refresh_token")
+        if refresh_token is None:
+            return Response({'error': 'No refresh'}, status=status.HTTP_400_BAD_REQUEST)
+
+        # リクエストにリフレッシュトークンを含めなおす
+        request.data["refresh"] = refresh_token
+
         serializer = self.get_serializer(data=request.data)
 
         try:
@@ -61,11 +65,41 @@ class JWTokenRefreshView(views.TokenRefreshView):
         
         response = Response(serializer.validated_data, status=status.HTTP_200_OK)
 
-        response.delete_cookie("access_token")
-
         response.set_cookie(
             "access_token",
             serializer.validated_data["access"],
-            max_age=60 * 60,
+            max_age=60 * 60 * 3,
             httponly=True,
         )
+        response.set_cookie(
+            "refresh_token",
+            serializer.validated_data["refresh"],
+            max_age=60 * 60 * 24 * 7,
+            httponly=True,
+        )
+
+        return response
+
+class LogoutView(views.TokenBlacklistView):
+    authentication_classes = (CookieJWTAuthentication,)
+    permission_classes = (permissions.IsAuthenticated,)
+    def post(self, request, *args, **kwargs):
+        refresh_token = request.COOKIES.get("refresh_token")
+        if refresh_token is None:
+            return Response({'error': 'No refresh'}, status=status.HTTP_400_BAD_REQUEST)
+
+        # リクエストにリフレッシュトークンを含めなおす
+        request.data["refresh"] = refresh_token
+
+        response = super().post(request, *args, **kwargs)
+
+        # トークンをCookieから削除
+        response.delete_cookie("access_token")
+        response.delete_cookie("refresh_token")
+
+        # 既に存在するresponseにdataを追加
+        response.data = {"Message": "Logout"}
+
+        return response
+
+
